+++ tests/unit/test_users.py
@@ -10,16 +10,16 @@ from shushu import users
 
 def test_current_returns_current_user_info():
     info = users.current()
-    expected_name = pwd.getpwuid(os.getuid()).pw_name
+    expected_name = pwd.getpwuid(os.geteuid()).pw_name
     assert info.name == expected_name
-    assert info.uid == os.getuid()
+    assert info.uid == os.geteuid()
 
 
 def test_resolve_known_user_returns_info():
-    expected_name = pwd.getpwuid(os.getuid()).pw_name
+    expected_name = pwd.getpwuid(os.geteuid()).pw_name
     info = users.resolve(expected_name)
     assert info.name == expected_name
-    assert info.uid == os.getuid()
+    assert info.uid == os.geteuid()
 
 
 def test_resolve_unknown_user_raises():